Optimization results with multiple support states
UNSTABLE: After reviewing the code, almost everything seems to be implemented properly EXCEPT for the fact that in my implementation, increasing N also changes the total duration of the trajectory. Therefore, the best option is to modify the code and remove the assumption that each support state is separated by 1 second.