Added 2 test files and SDF basis
SDF (Signed Distance Field) will provide the interface for adding obstacles and computing gradients
Test2 represents a trajectory formed by mapping (a, b) to (0, b-a). This implementation has issues and causes ridges in the mean prior.
Test3 represents a true trajectory between (a, b), but lacks all the previous code. Test3 should not be modified further, but rather combined with the other tests to form the correct implementation.